requirements.txt
pydantic
python-dateutil
pandas
openpyxl
tabulate
PyQt6
approvaltests
pytest-approvaltests
